refactor(eve): split eval session content helpers
